refactor(secrets): split runtime activation helpers
